requirements.txt
openai
streamlit
langchain
python-dotenv
docx2txt
pypdf
